feat(runner): centralize coord-mcp provisioning at the terminal chokepoint
Move device-JWT coord-mcp provisioning out of the per-terminal call sites
into the one function every device-identity terminal funnels through
(acquire_for_terminal), so all six device-JWT terminal paths (operator tab,
HTTP /terminals, tauri proxy, backend relay, coordinator/worker sessions)
get a coord-mcp .mcp.json in one place instead of two.

- Extract provision_coord_mcp_for_session, write_coord_mcp_config,
  coord_mcp_safe_to_write, jwt_unverified_claim into a new leaf module
  coord_mcp.rs (with its own coord_http_base copy) — breaks the would-be
  agent_worktree -> agent_runtime -> agent_worktree cycle.
- Provision once at the end of acquire_for_terminal on the post-acquire cwd,
  covering every return path via a single call.
- Delete the now-redundant explicit call #529 added to terminal_create.
- Guard coord_mcp_safe_to_write so a device JWT never downgrades an existing
  agent-JWT coord-mcp config (decode the existing bearer's sub_type).

Closes the runner-spawned-session remainder from
plans/2026-06-09-centralize-coord-mcp-provisioning-terminal-chokepoint.md

Session-Id: a632551e-6c5f-4c58-99bf-46665ef0180a
