+++ src-tauri/src/coord_mcp.rs
@@ -1,0 +1,336 @@
+//! Coord-mcp `.mcp.json` provisioning — the leaf module owning every helper that
+//! writes a coord device/agent JWT into a session's `.mcp.json`.
+//!
+//! Extracted out of `agent_runtime.rs` so the terminal chokepoint
+//! (`agent_worktree::isolated_edit::acquire_for_terminal`) can call provisioning
+//! WITHOUT creating a circular dependency: `agent_runtime` already depends on
+//! `agent_worktree`, so `agent_worktree → agent_runtime` would close a cycle.
+//! This module depends only on `crate::auth`, `std::fs`, `serde_json`, `base64`
+//! — a true leaf both callers can share.
+
+use std::path::Path;
+use tracing::{info, warn};
+
+/// Resolve the coord HTTP base: `COORD_HTTP_URL` env first, then the active
+/// profile's `coord_url` (ws→http normalized). `None` when neither is set.
+/// Inline copy of `agent_runtime::coord_http_base` — kept local so this leaf
+/// module does NOT depend on `agent_runtime` (which depends back on us).
+fn coord_http_base() -> Option<String> {
+    if let Ok(v) = std::env::var("COORD_HTTP_URL") {
+        if !v.is_empty() {
+            return Some(v);
+        }
+    }
+    let coord_url = qontinui_runner_lib::profiles::load_strict()
+        .ok()?
+        .coord_url?;
+    let trimmed = coord_url.trim_end_matches("/ws");
+    let with_http = trimmed
+        .strip_prefix("wss://")
+        .map(|rest| format!("https://{rest}"))
+        .or_else(|| {
+            trimmed
+                .strip_prefix("ws://")
+                .map(|rest| format!("http://{rest}"))
+        })
+        .unwrap_or_else(|| trimmed.to_string());
+    Some(with_http)
+}
+
+/// The coord-native `/mcp` endpoint is live (coord PR #277 Phase-2 cutover),
+/// so this no longer carries the prior "do not deploy" gate. If a target coord
+/// ever lacks the `/mcp` route, agents get an unreachable MCP server: Claude
+/// Code degrades gracefully and runs *without* coord tools (a silent
+/// coordination regression) — so always sequence a coord `/mcp` deploy ahead
+/// of pointing runners at a new coord.
+///
+/// Coord base resolution: `COORD_HTTP_URL` env → active profile's `coord_url`
+/// → localhost fallback. Previously this read ONLY the env var, so a
+/// profile-configured runner (production → `coord.qontinui.io`) wrote a
+/// `localhost` MCP url into the agent's `.mcp.json`, silently pointing spawned
+/// agents at the wrong coord.
+pub(crate) fn write_coord_mcp_config(primary_wt: &str, jwt: &str) {
+    let coord_url = std::env::var("COORD_HTTP_URL")
+        .ok()
+        .filter(|v| !v.trim().is_empty())
+        .or_else(coord_http_base)
+        .unwrap_or_else(|| "http://localhost:9870".to_string());
+    let mcp_url = format!("{}/mcp", coord_url.trim_end_matches('/'));
+
+    let mcp_config = serde_json::json!({
+        "mcpServers": {
+            "coord-mcp": {
+                "type": "http",
+                "url": mcp_url,
+                "headers": {
+                    "Authorization": format!("Bearer {}", jwt),
+                }
+            }
+        }
+    });
+
+    let mcp_path = Path::new(primary_wt).join(".mcp.json");
+    match std::fs::write(
+        &mcp_path,
+        serde_json::to_string_pretty(&mcp_config).unwrap_or_default(),
+    ) {
+        Ok(()) => {
+            info!(
+                "agent_runtime: wrote .mcp.json for coord-mcp in {}",
+                primary_wt
+            );
+        }
+        Err(e) => {
+            warn!(
+                "agent_runtime: failed to write .mcp.json in {}: {e}",
+                primary_wt
+            );
+        }
+    }
+}
+
+/// Provision `.mcp.json` for a runner-spawned session that did NOT arrive with a
+/// coord-minted agent JWT — i.e. gate-continuation terminals. Uses the runner's
+/// own live **device** JWT (the same `SubType::Device` EdDSA token the coord
+/// `/mcp` relay presents) so the continuation can use `coord_register_gate` over
+/// MCP and `coord-acting-bearer.sh` (which mints an acting-user Service token
+/// from it). This closes the reach gap where continuations — a primary place
+/// follow-up gates get registered — otherwise fell back to the operator-bearer
+/// stopgap (plan 2026-06-09-provision-coord-mcp-on-all-runner-spawned-sessions).
+///
+/// Two guards the agent-spawn path does not need (it always writes a coord agent
+/// JWT into a fresh worktree):
+/// 1. **sub_type guard.** The `access_token` slot can hold a *Cognito* token
+///    rather than the coord device JWT on some pairing tiers
+///    (`device_jwt_refresher`); a Cognito bearer would 401 against coord's EdDSA
+///    verifier. Write only when the bearer decodes as `sub_type ∈ {device,
+///    agent}`; otherwise log + skip (never write a non-verifying bearer).
+/// 2. **non-clobber guard.** A continuation can degrade to a canonical repo
+///    checkout (worktree mode off / acquire declined), which may already hold
+///    the operator's own `.mcp.json`. Overwrite only when the file is absent or
+///    is solely our `coord-mcp` config (see [`coord_mcp_safe_to_write`]).
+// `pub(crate)` so the operator-opened-tab entry point
+// (`commands::terminal::terminal_create`) reuses this exact helper — closing the
+// last dark runner-spawned session kind (plan
+// 2026-06-09-provision-coord-mcp-operator-tabs). Both modules compile into the
+// runner bin crate, so `pub(crate)` is the minimal visibility that reaches it.
+pub(crate) fn provision_coord_mcp_for_session(workdir: &str) {
+    let jwt = match crate::auth::AuthManager::new().get_access_token() {
+        Ok(t) if !t.trim().is_empty() => t,
+        _ => {
+            info!(
+                "agent_runtime: no device JWT in access_token slot — skipping \
+                 coord-mcp provisioning for {workdir}"
+            );
+            return;
+        }
+    };
+
+    match jwt_unverified_claim(&jwt, "sub_type").as_deref() {
+        Some("device") | Some("agent") => {}
+        other => {
+            info!(
+                "agent_runtime: access_token bearer is not a coord device/agent \
+                 JWT (sub_type={other:?}) — skipping coord-mcp provisioning for \
+                 {workdir} (would 401 against coord's EdDSA verifier)"
+            );
+            return;
+        }
+    }
+
+    if !coord_mcp_safe_to_write(workdir) {
+        info!(
+            "agent_runtime: {workdir}/.mcp.json already holds a non-coord-mcp \
+             config — leaving it untouched (no coord-mcp provisioning)"
+        );
+        return;
+    }
+
+    write_coord_mcp_config(workdir, &jwt);
+}
+
+/// Decode an unverified string claim from a JWT payload. We do NOT verify the
+/// signature — coord re-validates the EdDSA signature on use; here we only need
+/// `sub_type` to avoid writing a non-coord-verifying bearer (e.g. a Cognito
+/// token) into `.mcp.json`. Mirror of `cognito::jwt_claim`, inlined because that
+/// fn lives in the lib crate (`pub(crate)`) and `agent_runtime` compiles into
+/// the bin crate, which does not re-declare `mod cognito`.
+fn jwt_unverified_claim(token: &str, claim: &str) -> Option<String> {
+    use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine as _};
+    let mut parts = token.splitn(3, '.');
+    let _header = parts.next()?;
+    let payload_b64 = parts.next()?;
+    let _sig = parts.next()?;
+    let payload = URL_SAFE_NO_PAD.decode(payload_b64).ok()?;
+    let json: serde_json::Value = serde_json::from_slice(&payload).ok()?;
+    json.get(claim).and_then(|v| v.as_str()).map(String::from)
+}
+
+/// True iff writing our coord-mcp `.mcp.json` into `workdir` would not clobber a
+/// user's own config: the file is absent/unreadable, OR it parses as a config
+/// whose `mcpServers` is solely our `coord-mcp` entry (a prior provisioning we
+/// own and may refresh). A foreign or unparseable file returns false (leave it).
+fn coord_mcp_safe_to_write(workdir: &str) -> bool {
+    let path = Path::new(workdir).join(".mcp.json");
+    let existing = match std::fs::read_to_string(&path) {
+        Ok(s) => s,
+        Err(_) => return true, // absent (or unreadable) → safe to create
+    };
+    let parsed: serde_json::Value = match serde_json::from_str(&existing) {
+        Ok(v) => v,
+        Err(_) => return false, // unparseable foreign file → do not clobber
+    };
+    match parsed.get("mcpServers").and_then(|m| m.as_object()) {
+        Some(servers) => {
+            if servers.len() == 1 && servers.contains_key("coord-mcp") {
+                // Our own coord-mcp config — refreshable, EXCEPT never downgrade an
+                // existing agent JWT (richer scopes) to a device JWT. If the current
+                // bearer decodes sub_type=agent, leave it.
+                let existing_is_agent = parsed
+                    .pointer("/mcpServers/coord-mcp/headers/Authorization")
+                    .and_then(|v| v.as_str())
+                    .and_then(|h| h.strip_prefix("Bearer "))
+                    .and_then(|tok| jwt_unverified_claim(tok, "sub_type"))
+                    .map(|st| st == "agent")
+                    .unwrap_or(false);
+                !existing_is_agent
+            } else {
+                false
+            }
+        }
+        None => false,
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn write_coord_mcp_config_emits_http_bearer_shape() {
+        let tmp = std::env::temp_dir().join(format!("coord-mcp-cfg-{}", uuid::Uuid::now_v7()));
+        std::fs::create_dir_all(&tmp).unwrap();
+        let primary_wt = tmp.to_string_lossy().to_string();
+
+        let prev = std::env::var("COORD_HTTP_URL").ok();
+        std::env::set_var("COORD_HTTP_URL", "https://coord.example.test/");
+
+        write_coord_mcp_config(&primary_wt, "header.payload.sig");
+
+        let written = std::fs::read_to_string(tmp.join(".mcp.json")).unwrap();
+        let v: serde_json::Value = serde_json::from_str(&written).unwrap();
+        let server = &v["mcpServers"]["coord-mcp"];
+
+        // HTTP transport pointing at coord /mcp, Bearer-authenticated.
+        assert_eq!(server["type"], "http");
+        assert_eq!(server["url"], "https://coord.example.test/mcp");
+        assert_eq!(
+            server["headers"]["Authorization"],
+            "Bearer header.payload.sig"
+        );
+
+        // No Node-sidecar/subprocess residue, and no identity env vars
+        // (identity is derived server-side from the JWT claims).
+        assert!(server.get("command").is_none(), "must not spawn a command");
+        assert!(server.get("args").is_none(), "must not pass node args");
+        assert!(server.get("env").is_none(), "identity must come from JWT");
+        assert!(
+            !written.contains("node") && !written.contains("coord-mcp.mjs"),
+            "config must not reference the Node sidecar: {written}"
+        );
+
+        // Cleanup.
+        let _ = std::fs::remove_dir_all(&tmp);
+        match prev {
+            Some(p) => std::env::set_var("COORD_HTTP_URL", p),
+            None => std::env::remove_var("COORD_HTTP_URL"),
+        }
+    }
+
+    /// `coord_mcp_safe_to_write` — the non-clobber guard for the continuation
+    /// provisioning path (continuations can degrade to a real repo checkout that
+    /// already holds the operator's own `.mcp.json`).
+    #[test]
+    fn coord_mcp_safe_to_write_guards_user_config() {
+        let dir = std::env::temp_dir().join(format!("coord-mcp-safe-{}", uuid::Uuid::now_v7()));
+        std::fs::create_dir_all(&dir).unwrap();
+        let wd = dir.to_string_lossy().to_string();
+        let mcp = dir.join(".mcp.json");
+
+        // 1. Absent → safe to create.
+        assert!(coord_mcp_safe_to_write(&wd), "absent file must be writable");
+
+        // 2. Solely our coord-mcp config → safe to refresh (we own it).
+        std::fs::write(
+            &mcp,
+            r#"{"mcpServers":{"coord-mcp":{"type":"http","url":"https://c/mcp"}}}"#,
+        )
+        .unwrap();
+        assert!(
+            coord_mcp_safe_to_write(&wd),
+            "a solely-coord-mcp config is ours — refreshable"
+        );
+
+        // 3. A user's own config (different server) → must NOT clobber.
+        std::fs::write(
+            &mcp,
+            r#"{"mcpServers":{"my-server":{"type":"http","url":"https://x/mcp"}}}"#,
+        )
+        .unwrap();
+        assert!(
+            !coord_mcp_safe_to_write(&wd),
+            "a foreign mcpServers config must be left untouched"
+        );
+
+        // 4. coord-mcp ALONGSIDE another server (2 keys) → not solely ours → skip.
+        std::fs::write(
+            &mcp,
+            r#"{"mcpServers":{"coord-mcp":{"url":"https://c/mcp"},"other":{"url":"x"}}}"#,
+        )
+        .unwrap();
+        assert!(
+            !coord_mcp_safe_to_write(&wd),
+            "coord-mcp plus a user server is the user's file — do not clobber"
+        );
+
+        // 5. Unparseable / non-JSON → conservatively do not clobber.
+        std::fs::write(&mcp, "not json {{{").unwrap();
+        assert!(
+            !coord_mcp_safe_to_write(&wd),
+            "an unparseable file must be left untouched"
+        );
+
+        // 6. Our coord-mcp config but bearer is an AGENT JWT → must NOT downgrade.
+        use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine as _};
+        let agent_payload = URL_SAFE_NO_PAD.encode(br#"{"sub_type":"agent"}"#);
+        let agent_jwt = format!("h.{agent_payload}.s");
+        std::fs::write(
+            &mcp,
+            format!(
+                r#"{{"mcpServers":{{"coord-mcp":{{"type":"http","url":"https://c/mcp","headers":{{"Authorization":"Bearer {agent_jwt}"}}}}}}}}"#
+            ),
+        )
+        .unwrap();
+        assert!(
+            !coord_mcp_safe_to_write(&wd),
+            "must not downgrade an existing agent-JWT coord-mcp config to a device JWT"
+        );
+
+        // 7. Same shape but a DEVICE-JWT bearer → still refreshable (ours, same tier).
+        let device_payload = URL_SAFE_NO_PAD.encode(br#"{"sub_type":"device"}"#);
+        let device_jwt = format!("h.{device_payload}.s");
+        std::fs::write(
+            &mcp,
+            format!(
+                r#"{{"mcpServers":{{"coord-mcp":{{"type":"http","url":"https://c/mcp","headers":{{"Authorization":"Bearer {device_jwt}"}}}}}}}}"#
+            ),
+        )
+        .unwrap();
+        assert!(
+            coord_mcp_safe_to_write(&wd),
+            "a device-JWT coord-mcp config is ours at the same tier — refreshable"
+        );
+
+        let _ = std::fs::remove_dir_all(&dir);
+    }
+}
